+++ src/server.ts
@@ -2,7 +2,6 @@ import * as WebSocket from 'ws';
 
 import MessageTypes from './message-types';
 import { GRAPHQL_WS, GRAPHQL_SUBSCRIPTIONS } from './protocol';
-import { SubscriptionManager } from 'graphql-subscriptions';
 import isObject = require('lodash.isobject');
 import {
   parse,
@@ -13,7 +12,6 @@ import {
   ValidationContext,
   specifiedRules,
 } from 'graphql';
-import { executeFromSubscriptionManager } from './adapters/subscription-manager';
 import { createEmptyIterable } from './utils/empty-iterable';
 import { createAsyncIterator, forAwaitEach, isAsyncIterable } from 'iterall';
 import { createIterableFromPromise } from './utils/promise-to-iterable';
@@ -75,10 +73,6 @@ export interface ServerOptions {
   onDisconnect?: Function;
   keepAlive?: number;
 
-  /**
-   * @deprecated subscriptionManager is deprecated, use executor instead
-   */
-  subscriptionManager?: SubscriptionManager;
   /**
    * @deprecated onSubscribe is deprecated, use onOperation instead
    */
@@ -213,37 +207,20 @@ export class SubscriptionServer {
   }
 
   private loadExecutor(options: ServerOptions) {
-    const { subscriptionManager, execute, subscribe, schema, rootValue } = options;
+    const { execute, subscribe, schema, rootValue } = options;
 
-    if (!subscriptionManager && !execute) {
-      throw new Error('Must provide `subscriptionManager` or `execute` to websocket server constructor.');
+    if (!execute) {
+      throw new Error('Must provide `execute` for websocket server constructor.');
     }
 
-    if (subscriptionManager && execute) {
-      throw new Error('Must provide `subscriptionManager` or `execute` and not both.');
-    }
-
-    if (subscribe && !execute) {
-      throw new Error('Must provide `execute` when providing `subscribe`!');
-    }
-
-    if (execute && !schema) {
-      throw new Error('Must provide `schema` when using `execute`.');
-    }
-
-    if (subscriptionManager) {
-      console.warn('subscriptionManager is deprecated, use `execute` or `subscribe` directly from `graphql-js`!');
+    if (!schema) {
+      throw new Error('`schema` is missing');
     }
 
     this.schema = schema;
     this.rootValue = rootValue;
-
-    if (subscriptionManager) {
-      this.execute = executeFromSubscriptionManager(subscriptionManager);
-    } else {
-      this.execute = execute;
-      this.subscribe = subscribe;
-    }
+    this.execute = execute;
+    this.subscribe = subscribe;
   }
 
   private unsubscribe(connectionContext: ConnectionContext, opId: string) {
@@ -388,14 +365,7 @@ export class SubscriptionServer {
 
               const document = typeof baseParams.query !== 'string' ? baseParams.query : parse(baseParams.query);
               let executionIterable: AsyncIterator<ExecutionResult> | Promise<AsyncIterator<ExecutionResult>>;
-              let validationErrors: Error[] = [];
-
-              if ( this.schema ) {
-                // NOTE: This is a temporary condition to support the legacy subscription manager.
-                // As soon as subscriptionManager support is removed, we can remove the if
-                // and keep only the validation part.
-                validationErrors = validate(this.schema, document, this.specifiedRules);
-              }
+              const validationErrors: Error[] = validate(this.schema, document, this.specifiedRules);
 
               if ( validationErrors.length > 0 ) {
                 executionIterable = createIterableFromPromise<ExecutionResult>(